requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.5
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dominate==2.5.1
entrypoints==0.3
future==0.18.2
h5py==2.9.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.13.3
Jinja2==2.11.2
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.6
nose==1.3.7
notebook==6.0.3
numpy==1.16.3
opencv-python==4.2.0.34
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.11.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
pyparsing==2.4.0
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.16.0
pytest==4.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sympy==1.4
terminado==0.8.3
testpath==0.4.4
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.2
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
zipp==3.1.0
